pl1.x, with bug

+++ model.py
@@ -178,8 +178,10 @@ class Slice(nn.Module):
         if device >= 0:
             hg = hg.to(device)
             wg = wg.to(device)
-        hg = hg.float().repeat(N, 1, 1).unsqueeze(3) / (H-1) * 2 - 1 # norm to [-1,1] NxHxWx1
-        wg = wg.float().repeat(N, 1, 1).unsqueeze(3) / (W-1) * 2 - 1 # norm to [-1,1] NxHxWx1
+        # hg = hg.float().repeat(N, 1, 1).unsqueeze(3) / (H-1) * 2 - 1 # norm to [-1,1] NxHxWx1
+        # wg = wg.float().repeat(N, 1, 1).unsqueeze(3) / (W-1) * 2 - 1 # norm to [-1,1] NxHxWx1
+        hg = hg.type_as(bilateral_grid).repeat(N, 1, 1).unsqueeze(3) / (H-1) * 2 - 1 # norm to [-1,1] NxHxWx1
+        wg = wg.type_as(bilateral_grid).repeat(N, 1, 1).unsqueeze(3) / (W-1) * 2 - 1 # norm to [-1,1] NxHxWx1
         guidemap = guidemap.permute(0,2,3,1).contiguous()
         guidemap_guide = torch.cat([wg, hg, guidemap], dim=3).unsqueeze(1) # Nx1xHxWx3
         coeff = F.grid_sample(bilateral_grid, guidemap_guide, align_corners=True)
